requirements.txt
akshare
google-genai
tavily-python
python-dotenv
pandas
openai